+++ src/test/java/com/glasbergen/neat/examples/connectfour/BoardTest.java
@@ -10,19 +10,19 @@ public class BoardTest {
 	public void simpleBoardAdd(){
 		Board b = new Board();
 		b.addToBoard(0, Colour.BLACK);
-		int[] inputVec = b.boardStateToInputVector(Colour.BLACK);
+		double[] inputVec = b.boardStateToInputVector(Colour.BLACK);
 		for(int i = 0; i < 6*7*3; i+=3){
 			if( i == 105 ) { //bottom left position
-				assertThat(inputVec[i], equalTo(1));
-				assertThat(inputVec[i+1], equalTo(0));
-				assertThat(inputVec[i+2], equalTo(0));
+				assertThat((int)inputVec[i], equalTo(1));
+				assertThat((int)inputVec[i+1], equalTo(0));
+				assertThat((int)inputVec[i+2], equalTo(0));
 			} else {
-				assertThat(inputVec[i], equalTo(0));
-				assertThat(inputVec[i+1], equalTo(0));
-				assertThat(inputVec[i+2], equalTo(1));
+				assertThat((int)inputVec[i], equalTo(0));
+				assertThat((int)inputVec[i+1], equalTo(0));
+				assertThat((int)inputVec[i+2], equalTo(1));
 			}
 		}
-		assertThat(inputVec[6*7*3], equalTo(1));
+		assertThat((int)inputVec[6*7*3], equalTo(1));
 	}
 	
 	@Test
@@ -32,16 +32,16 @@ public class BoardTest {
 		b.addToBoard(1, Colour.BLACK);
 		b.addToBoard(2, Colour.BLACK);
 		assertThat(b.addToBoard(3, Colour.BLACK), equalTo(true));
-		int[] inputVec = b.boardStateToInputVector(Colour.BLACK);
+		double[] inputVec = b.boardStateToInputVector(Colour.BLACK);
 		for(int i = 0; i < 6*7*3; i+=3){
 			if( i >= 105 && i <= 114) { //bottom left 4 positions
-				assertThat(inputVec[i], equalTo(1));
-				assertThat(inputVec[i+1], equalTo(0));
-				assertThat(inputVec[i+2], equalTo(0));
+				assertThat((int)inputVec[i], equalTo(1));
+				assertThat((int)inputVec[i+1], equalTo(0));
+				assertThat((int)inputVec[i+2], equalTo(0));
 			} else {
-				assertThat(inputVec[i], equalTo(0));
-				assertThat(inputVec[i+1], equalTo(0));
-				assertThat(inputVec[i+2], equalTo(1));
+				assertThat((int)inputVec[i], equalTo(0));
+				assertThat((int)inputVec[i+1], equalTo(0));
+				assertThat((int)inputVec[i+2], equalTo(1));
 			}
 		}
 	}
@@ -53,16 +53,16 @@ public class BoardTest {
 		b.addToBoard(5, Colour.BLACK);
 		b.addToBoard(4, Colour.BLACK);
 		assertThat(b.addToBoard(3, Colour.BLACK), equalTo(true));
-		int[] inputVec = b.boardStateToInputVector(Colour.BLACK);
+		double[] inputVec = b.boardStateToInputVector(Colour.BLACK);
 		for(int i = 0; i < 6*7*3; i+=3){
 			if( i >= 114 && i <= 123) { //bottom right 4 positions
-				assertThat(inputVec[i], equalTo(1));
-				assertThat(inputVec[i+1], equalTo(0));
-				assertThat(inputVec[i+2], equalTo(0));
+				assertThat((int)inputVec[i], equalTo(1));
+				assertThat((int)inputVec[i+1], equalTo(0));
+				assertThat((int)inputVec[i+2], equalTo(0));
 			} else {
-				assertThat(inputVec[i], equalTo(0));
-				assertThat(inputVec[i+1], equalTo(0));
-				assertThat(inputVec[i+2], equalTo(1));
+				assertThat((int)inputVec[i], equalTo(0));
+				assertThat((int)inputVec[i+1], equalTo(0));
+				assertThat((int)inputVec[i+2], equalTo(1));
 			}
 		}
 	}
@@ -74,16 +74,16 @@ public class BoardTest {
 		b.addToBoard(0, Colour.BLACK);
 		b.addToBoard(0, Colour.BLACK);
 		assertThat(b.addToBoard(0, Colour.BLACK), equalTo(true));
-		int[] inputVec = b.boardStateToInputVector(Colour.BLACK);
+		double[] inputVec = b.boardStateToInputVector(Colour.BLACK);
 		for(int i = 0; i < 6*7*3; i+=3){
 			if( i == 42 || i == 63 || i ==  84 || i == 105) { //stack bottom left
-				assertThat(inputVec[i], equalTo(1));
-				assertThat(inputVec[i+1], equalTo(0));
-				assertThat(inputVec[i+2], equalTo(0));
+				assertThat((int)inputVec[i], equalTo(1));
+				assertThat((int)inputVec[i+1], equalTo(0));
+				assertThat((int)inputVec[i+2], equalTo(0));
 			} else {
-				assertThat(inputVec[i], equalTo(0));
-				assertThat(inputVec[i+1], equalTo(0));
-				assertThat(inputVec[i+2], equalTo(1));
+				assertThat((int)inputVec[i], equalTo(0));
+				assertThat((int)inputVec[i+1], equalTo(0));
+				assertThat((int)inputVec[i+2], equalTo(1));
 			}
 		}
 	}
@@ -95,36 +95,36 @@ public class BoardTest {
 		b.addToBoard(0, Colour.RED);
 		b.addToBoard(0, Colour.BLACK);
 		assertThat(b.addToBoard(0, Colour.BLACK), equalTo(false));
-		int[] inputVec = b.boardStateToInputVector(Colour.BLACK);
+		double[] inputVec = b.boardStateToInputVector(Colour.BLACK);
 		for(int i = 0; i < 6*7*3; i+=3){
 			if( i == 42 || i == 63 || i == 105) { //stack bottom left
-				assertThat(inputVec[i], equalTo(1));
-				assertThat(inputVec[i+1], equalTo(0));
-				assertThat(inputVec[i+2], equalTo(0));
+				assertThat((int)inputVec[i], equalTo(1));
+				assertThat((int)inputVec[i+1], equalTo(0));
+				assertThat((int)inputVec[i+2], equalTo(0));
 			} else if( i == 84 ) {
-				assertThat(inputVec[i], equalTo(0));
-				assertThat(inputVec[i+1], equalTo(1));
-				assertThat(inputVec[i+2], equalTo(0));
+				assertThat((int)inputVec[i], equalTo(0));
+				assertThat((int)inputVec[i+1], equalTo(1));
+				assertThat((int)inputVec[i+2], equalTo(0));
 			} else {
-				assertThat(inputVec[i], equalTo(0));
-				assertThat(inputVec[i+1], equalTo(0));
-				assertThat(inputVec[i+2], equalTo(1));
+				assertThat((int)inputVec[i], equalTo(0));
+				assertThat((int)inputVec[i+1], equalTo(0));
+				assertThat((int)inputVec[i+2], equalTo(1));
 			}
 		}
 		inputVec = b.boardStateToInputVector(Colour.RED);
 		for(int i = 0; i < 6*7*3; i+=3){
 			if( i == 42 || i == 63 || i == 105) { //stack bottom left
-				assertThat(inputVec[i], equalTo(0));
-				assertThat(inputVec[i+1], equalTo(1));
-				assertThat(inputVec[i+2], equalTo(0));
+				assertThat((int)inputVec[i], equalTo(0));
+				assertThat((int)inputVec[i+1], equalTo(1));
+				assertThat((int)inputVec[i+2], equalTo(0));
 			} else if( i == 84 ) {
-				assertThat(inputVec[i], equalTo(1));
-				assertThat(inputVec[i+1], equalTo(0));
-				assertThat(inputVec[i+2], equalTo(0));
+				assertThat((int)inputVec[i], equalTo(1));
+				assertThat((int)inputVec[i+1], equalTo(0));
+				assertThat((int)inputVec[i+2], equalTo(0));
 			} else {
-				assertThat(inputVec[i], equalTo(0));
-				assertThat(inputVec[i+1], equalTo(0));
-				assertThat(inputVec[i+2], equalTo(1));
+				assertThat((int)inputVec[i], equalTo(0));
+				assertThat((int)inputVec[i+1], equalTo(0));
+				assertThat((int)inputVec[i+2], equalTo(1));
 			}
 		}
 	}
